requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
pgvector==0.3.2

# Pydantic & settings
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1

# Supabase / HTTP
supabase==2.4.6
httpx==0.27.0

# Google Drive (audio file storage — conserves Supabase free-tier quota)
google-api-python-client>=2.130.0
google-auth>=2.29.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Audio processing
librosa==0.10.2
soundfile==0.12.1
numpy>=1.26.0  # TF 2.20 requires >=1.26; _utils module added in 1.25 (was missing in 1.23)

# ML models (CLAP)
torch==2.3.0
torchvision==0.18.0
transformers==4.41.2
laion-clap==1.1.4

# YAMNet + MusiCNN (TensorFlow ecosystem)
# TF 2.20 ships with Keras 3 which removed tf.compat.v1.layers.
# tf-keras restores the Keras 2 / legacy-layer support required by musicnn.
# Set TF_USE_LEGACY_KERAS=1 before importing TF (done in app/workers/registry.py).
tensorflow==2.20.0
tensorflow-hub==0.16.0
tf-keras==2.20.1

# setuptools<71 required: tensorflow-hub 0.16 uses `from pkg_resources import parse_version`
# which was removed in setuptools 71+. Pin to <=70 to keep pkg_resources importable.
setuptools<71

# MusiCNN — MTG music tagger (genre, mood, instrumentation)
# Install with --no-deps: musicnn pins numpy<1.17 which conflicts with TF 2.20.
# musicnn==0.1.0   →  pip install --no-deps musicnn==0.1.0
